+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 # Unreleased
 
+## [1.1.0] : 2019-04-04
+
++ [Fix] Now it will turn other text layer off.  
+ But the text layer should'nt be export.
++ [Feature] It it found same name text layer in a group,  
+ it will export with layer at same time.
++ [Feature] Now it will search all text layer's contexts,  
+ and write to txt file at current psd file folder.
+
 ## [1.0.0] : 2019-01-13
 
 + First build.